Move some tests from db_test(2) to compression_test
Summary: ... to improve compilation times on db_test and db_test2 and to
consolidate more compression-related tests into compression_test.

Test Plan: existing tests, and seems like I haven't thrown anything
away:
```
$ git diff | grep -Ec '^[-]' # lines removed
1535
$ git diff | grep -Ec '^[+]' # lines added
1535
$
```
